db_api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary
python-dotenv==1.0.0